Rename Climate.h/cpp to Weather.h/cpp
