+++ src/openrct2/Cheats.cpp
@@ -16,7 +16,7 @@
 #include "core/DataSerialiser.h"
 #include "network/Network.h"
 #include "ride/Ride.h"
-#include "world/Climate.h"
+#include "world/Weather.h"
 #include "world/Footpath.h"
 #include "world/Park.h"
 #include "world/Scenery.h"
