+++ src/openrct2/Game.cpp
@@ -66,7 +66,7 @@
 #include "ui/WindowManager.h"
 #include "windows/Intent.h"
 #include "world/Banner.h"
-#include "world/Climate.h"
+#include "world/Weather.h"
 #include "world/Entrance.h"
 #include "world/Footpath.h"
 #include "world/Map.h"
