+++ src/openrct2/audio/Audio.cpp
@@ -29,7 +29,7 @@
 #include "../scenes/intro/IntroScene.h"
 #include "../ui/WindowManager.h"
 #include "../util/Util.h"
-#include "../world/Climate.h"
+#include "../world/Weather.h"
 #include "../world/Map.h"
 #include "../world/tile_element/SurfaceElement.h"
 #include "AudioChannel.h"
