+++ src/openrct2/entity/Peep.cpp
@@ -51,7 +51,7 @@
 #include "../ui/WindowManager.h"
 #include "../util/Util.h"
 #include "../windows/Intent.h"
-#include "../world/Climate.h"
+#include "../world/Weather.h"
 #include "../world/ConstructionClearance.h"
 #include "../world/Entrance.h"
 #include "../world/Footpath.h"
