+++ src/openrct2/interface/Viewport.cpp
@@ -34,7 +34,7 @@
 #include "../ride/TrackDesign.h"
 #include "../ride/Vehicle.h"
 #include "../ui/WindowManager.h"
-#include "../world/Climate.h"
+#include "../world/Weather.h"
 #include "../world/Map.h"
 #include "../world/tile_element/LargeSceneryElement.h"
 #include "../world/tile_element/SmallSceneryElement.h"
