+++ src/openrct2/paint/Paint.Entity.cpp
@@ -24,7 +24,7 @@
 #include "../ride/RideData.h"
 #include "../ride/TrackDesign.h"
 #include "../ride/Vehicle.h"
-#include "../world/Climate.h"
+#include "../world/Weather.h"
 #include "../world/Map.h"
 #include "../world/Park.h"
 #include "Paint.h"
