+++ src/openrct2/ride/RideConstruction.cpp
@@ -26,7 +26,7 @@
 #include "../network/Network.h"
 #include "../ui/WindowManager.h"
 #include "../windows/Intent.h"
-#include "../world/Climate.h"
+#include "../world/Weather.h"
 #include "../world/Entrance.h"
 #include "../world/Footpath.h"
 #include "../world/Location.hpp"
