+++ src/openrct2/world/Weather.h
